refactor(component): ♻️  HuntParticipants as component
